Remove 3d Detection
