--- src/otx/algo/object_detection_3d/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Custom model implementations for object detection 3D task."""
-
-from . import backbones, detectors, heads, losses, matchers, utils
-
-__all__ = ["backbones", "heads", "losses", "detectors", "matchers", "utils"]
