--- src/otx/algo/object_detection_3d/detectors/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Core torch detectors modules for 3d object detection."""
