--- src/otx/algo/object_detection_3d/heads/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""heads modules for 3d object detection."""
-
-from .depth_predictor import DepthPredictor
-from .depthaware_transformer import DepthAwareTransformerBuilder
-
-__all__ = ["DepthPredictor", "DepthAwareTransformerBuilder"]
