--- src/otx/algo/object_detection_3d/losses/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Loss functions for 3d object detection."""
-
-from .monodetr_loss import MonoDETRCriterion
-
-__all__ = ["MonoDETRCriterion"]
