--- src/otx/algo/object_detection_3d/matchers/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Matchers modules for 3d object detection."""
-
-from .matcher_3d import HungarianMatcher3D
-
-__all__ = ["HungarianMatcher3D"]
