--- src/otx/algo/object_detection_3d/utils/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""utils module for object detection 3D models."""
