+++ src/otx/core/data/factory.py
@@ -67,7 +67,7 @@ class OTXDatasetFactory:
     """Factory class for OTXDataset."""
 
     @classmethod
-    def create(  # noqa: PLR0911
+    def create(
         cls: type[OTXDatasetFactory],
         task: OTXTaskType,
         dm_subset: DmDataset,
@@ -152,10 +152,5 @@ class OTXDatasetFactory:
             from .dataset.keypoint_detection import OTXKeypointDetectionDataset
 
             return OTXKeypointDetectionDataset(**common_kwargs)
-
-        if task == OTXTaskType.OBJECT_DETECTION_3D:
-            from .dataset.object_detection_3d import OTX3DObjectDetectionDataset
-
-            return OTX3DObjectDetectionDataset(**common_kwargs)
 
         raise NotImplementedError(task)
