+++ src/otx/core/types/task.py
@@ -28,7 +28,6 @@ class OTXTaskType(str, Enum):
     ROTATED_DETECTION = "ROTATED_DETECTION"
     DETECTION_SEMI_SL = "DETECTION_SEMI_SL"
     KEYPOINT_DETECTION = "KEYPOINT_DETECTION"
-    OBJECT_DETECTION_3D = "OBJECT_DETECTION_3D"
 
     # Segmentation
     INSTANCE_SEGMENTATION = "INSTANCE_SEGMENTATION"
