--- src/otx/recipe/object_detection_3d/monodetr3d.yaml
@@ -1,40 +1,0 @@
-model:
-  class_path: otx.algo.object_detection_3d.monodetr3d.MonoDETR3D
-  init_args:
-    label_info: 17
-    model_name: monodetr_50
-    input_size:
-      - 384
-      - 1280
-
-    optimizer:
-      class_path: torch.optim.AdamW
-      init_args:
-        lr: 0.0001
-        betas: [0.9, 0.999]
-        weight_decay: 0.0001
-
-    scheduler:
-      class_path: lightning.pytorch.cli.ReduceLROnPlateau
-      init_args:
-        mode: max
-        factor: 0.1
-        patience: 13
-        monitor: val/AP_2d@0.5
-
-engine:
-  task: OBJECT_DETECTION_3D
-  device: auto
-
-callback_monitor: val/AP_3d@0.5
-
-data: ../_base_/data/object_detection_3d.yaml
-
-precision: 32 # MonoDETR do not support fp16 training
-overrides:
-  callbacks:
-    - class_path: otx.algo.callbacks.adaptive_early_stopping.EarlyStoppingWithWarmup
-      init_args:
-        patience: 15
-        warmup_iters: 30
-        warmup_epochs: 30
