--- tests/perf/test_object_detection_3d.py
@@ -1,83 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-"""OTX 3d detection perfomance benchmark tests."""
-
-from __future__ import annotations
-
-from pathlib import Path
-from typing import ClassVar
-
-import pytest
-
-from .benchmark import Benchmark
-from .conftest import PerfTestBase
-
-
-class TestPerfObjectDetection3D(PerfTestBase):
-    """Benchmark visual prompting."""
-
-    MODEL_TEST_CASES = [  # noqa: RUF012
-        Benchmark.Model(task="object_detection_3d", name="monodetr3d", category="balance"),
-    ]
-
-    DATASET_TEST_CASES: ClassVar = [
-        Benchmark.Dataset(
-            name="kitti_medium_pedestrian_cyclist",
-            path=Path("object_detection_3d/medium_pedestrian_cyclist"),
-            group="medium",
-            num_repeat=5,
-            extra_overrides={},
-        ),
-        Benchmark.Dataset(
-            name="kitti_large_car",
-            path=Path("object_detection_3d/large_car"),
-            group="large",
-            num_repeat=5,
-            extra_overrides={},
-        ),
-    ]
-
-    BENCHMARK_CRITERIA = [  # noqa: RUF012
-        Benchmark.Criterion(name="train/epoch", summary="max", compare="<", margin=0.1),
-        Benchmark.Criterion(name="train/e2e_time", summary="max", compare="<", margin=0.1),
-        Benchmark.Criterion(name="val/AP_3d@0.5", summary="max", compare=">", margin=0.05),
-        Benchmark.Criterion(name="val/AP_2d@0.5", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="test/AP_3d@0.5", summary="max", compare=">", margin=0.05),
-        Benchmark.Criterion(name="test/AP_2d@0.5", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="export/AP_3d@0.5", summary="max", compare=">", margin=0.05),
-        Benchmark.Criterion(name="export/AP_2d@0.5", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="optimize/AP_3d@0.5", summary="max", compare=">", margin=0.05),
-        Benchmark.Criterion(name="optimize/AP_2d@0.5", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="train/iter_time", summary="mean", compare="<", margin=0.1),
-        Benchmark.Criterion(name="test/iter_time", summary="mean", compare="<", margin=0.1),
-        Benchmark.Criterion(name="export/iter_time", summary="mean", compare="<", margin=0.1),
-        Benchmark.Criterion(name="optimize/iter_time", summary="mean", compare="<", margin=0.1),
-        Benchmark.Criterion(name="test(train)/e2e_time", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="test(export)/e2e_time", summary="max", compare=">", margin=0.1),
-        Benchmark.Criterion(name="test(optimize)/e2e_time", summary="max", compare=">", margin=0.1),
-    ]
-
-    @pytest.mark.parametrize(
-        "fxt_model",
-        MODEL_TEST_CASES,
-        ids=lambda model: model.name,
-        indirect=True,
-    )
-    @pytest.mark.parametrize(
-        "fxt_dataset",
-        DATASET_TEST_CASES,
-        ids=lambda dataset: dataset.name,
-        indirect=True,
-    )
-    def test_perf(
-        self,
-        fxt_model: Benchmark.Model,
-        fxt_dataset: Benchmark.Dataset,
-        fxt_benchmark: Benchmark,
-    ):
-        self._test_perf(
-            model=fxt_model,
-            dataset=fxt_dataset,
-            benchmark=fxt_benchmark,
-            criteria=self.BENCHMARK_CRITERIA,
-        )
