--- tests/unit/algo/object_detection_3d/heads/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Test of OTX Object Detection 3D heads."""
